fixed friend search issues
